msm8953-common: Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Change-Id: I9a66ed6cd3550cc50d1dfd908528b1b3bfc4e4a2
Signed-off-by: rayrzy <vincentrayananda@gmail.com>